+++ src/utilities/error.ts
@@ -150,4 +150,12 @@ export enum Errors {
      * Raises when it is impossible to parse the statements.
      */
     parsingFailure = " Error parsing the statements.",
+    /**
+     * Raises when there is an issue restarting the LSP client.
+     */
+    lspRestartFailure = "Error restarting the LSP client.",
+    /**
+     * Raises when a fetch failes after a minute.
+     */
+    fetchTimeoutError = "Failed to fetch after a minute."
 }